optimizing with cython

+++ signatures/view.py
@@ -5,7 +5,7 @@ from utils.benchmark import benchmark
 @benchmark
 def expand_view_challenge_hash(params, digest: bytes, nb_views: int):
     prng = PRNG(params.security, digest)
-    views = np.zeros((params.nb_execs, nb_views), dtype=int)
+    views = np.zeros((params.nb_execs, nb_views), dtype=np.int32)
     
     for e in range(params.nb_execs):
         idx = 0
